refactor: Move ConfigModule to package configure.
