+++ src/main/java/org/opentripplanner/standalone/config/configure/ConfigModule.java
@@ -1,10 +1,15 @@
-package org.opentripplanner.standalone.config;
+package org.opentripplanner.standalone.config.configure;
 
 import dagger.Module;
 import dagger.Provides;
 import java.io.File;
 import javax.inject.Singleton;
 import org.opentripplanner.datastore.api.OtpDataStoreConfig;
+import org.opentripplanner.standalone.config.BuildConfig;
+import org.opentripplanner.standalone.config.ConfigLoader;
+import org.opentripplanner.standalone.config.ConfigModel;
+import org.opentripplanner.standalone.config.OtpConfig;
+import org.opentripplanner.standalone.config.RouterConfig;
 import org.opentripplanner.standalone.config.api.OtpBaseDirectory;
 
 @Module
